third_party/blink/web_tests/external/wpt/css/selectors/invalidation/state-in-has-expected.txt

This is a testharness.js-based test.
[FAIL] Test :has() invalidation with :state() pseudo-classes
  assert_equals: expected "rgb(0, 255, 0)" but got "rgb(255, 0, 0)"
Harness: the test ran to completion.
